Feature Dynamic-Form #7

+++ naps-frontend/src/app/auth/auth.service.ts
@@ -28,10 +28,12 @@ export class AuthService {
 
   logoutUser(): void {
     const refresh = this.localStorage.getRefresh();
-    this.api.delete(endpoints.login, refresh).subscribe(_ => {
-      this.localStorage.removeToken();
-      location.reload();
-    });
+    if (refresh) {
+      this.api.delete(endpoints.login, refresh).subscribe(_ => {
+        this.localStorage.removeToken();
+        location.reload();
+      });
+    }
   }
 
   isLoggedIn(): boolean {
